Update PollOption, User, PollAnswer to latest api specification

+++ include/tgbot/types/User.h
@@ -18,9 +18,12 @@ public:
     typedef std::shared_ptr<User> Ptr;
 
     /**
-     * @brief Unique identifier for this user or bot.
+     * @brief Unique identifier for this user or bot. This number may have more than 32 significant 
+     * bits and some programming languages may have difficulty/silent defects in interpreting it. 
+     * But it has at most 52 significant bits, so a 64-bit integer or double-precision float type 
+     * are safe for storing this identifier.
      */
-    std::int32_t id;
+    std::int64_t id;
 
     /**
      * @brief True, if this user is a bot
